refactor(entity): 将 Message 类中的 Long 类型字段改为 Integer 类型
- 将 id、senderId 和 receiverId 字段的类型从 Long 改为 Integer
- 这一变更可以减少内存消耗，提高数据处理效率

+++ backend/src/main/java/cn/hmck/entity/Message.java
@@ -11,11 +11,11 @@ import java.time.LocalDateTime;
 @TableName("er_message")
 public class Message implements Serializable {
     @TableId(type = IdType.AUTO)
-    private Long id;  // 消息ID
+    private Integer id;  // 消息ID
 
-    private Long senderId;  // 发送者ID
+    private Integer senderId;  // 发送者ID
 
-    private Long receiverId;  // 接收者ID
+    private Integer receiverId;  // 接收者ID
 
     private String content;  // 消息内容
